requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
coveralls==2.1.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.9
django-heroku==0.3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
